feat(localization): fluent prototype for i18n/l10n
This is a prototype using fluent for localization. It is based on the
simple app in the examples.

The main problem that I was having with fluent was if you want to make a module with some helper functions it becomes difficult due to the way you have to add a `FluentResource` to a `FluentBundle`.

There is a PR out to swap passing a reference of a `FluentResource` to a
`FluentBundle` and I have pinned the `fluent` package to that branch for
now. It made it much nicer because then you can wrap the
`FluentResource` in an `Arc` and it just works.

If we decide to move forward with this we should probably shuffle some
stuff around and write docs and all that, but wanted to get a prototype
out so folks could get an idea of what fluent looks like.